+++ src/admin/contact/verification/test_impl/test_email_exists.cc
@@ -34,7 +34,7 @@ namespace ContactVerification {
 
     FACTORY_MODULE_INIT_DEFI(TestEmailExists_init)
 
-    Test::T_run_result TestEmailExists::run(long _history_id) const {
+    Test::T_run_result TestEmailExists::run(unsigned long long _history_id) const {
         using std::string;
         using std::vector;
