+++ src/admin/contact/verification/test_impl/test_phone_syntax.h
@@ -48,7 +48,7 @@ namespace ContactVerification
                 // first draft of pattern - see ticket #9588
                 : PHONE_PATTERN ("^\\+[0-9]{1,3}\\.[0-9]{1,14}$") {}
 
-            virtual T_run_result run(long _history_id) const;
+            virtual T_run_result run(unsigned long long _history_id) const;
             static std::string registration_name() { return "phone_syntax"; }
     };
